// web/src/lib/openapi.d.ts
export interface components {
    schemas: {
        Error: {
            message: string;
        };
        HealthResponse: {
            /** @example ok */
            status: string;
            version: string;
        };
        UserDTO: {
            /** Format: int64 */
            id: number;
            email: string;
            role: string;
        };
        MeResponse: {
            user: components["schemas"]["UserDTO"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}